move RV stuff into model
